% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npx-is-available.R
\name{npx_is_available}
\alias{npx_is_available}
\alias{stop_if_npx_not_available}
\title{Check if 'npx' is available in the system PATH}
\usage{
npx_is_available()

stop_if_npx_not_available()
}
\value{
TRUE if 'npx' is available, FALSE otherwise
}
\description{
Check if 'npx' is available in the system PATH
}
\examples{
npx_is_available()
}
