% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw-init.R
\name{pw_init}
\alias{pw_init}
\title{Initialize Playwright}
\usage{
pw_init(
  where = golem::get_golem_wd(),
  create_playwright_version = "1.17.138",
  create_playwright_args = c("--no-examples", "--quiet")
)
}
\arguments{
\item{where}{The path to the package}

\item{create_playwright_version}{Version of create-playwright to use}

\item{create_playwright_args}{Additional arguments to pass to npx create-playwright}
}
\description{
This function initializes a Playwright project in the tests folder.
}
\examples{
\dontrun{
pw_init()
}
}
