% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_install.R
\name{pw_install}
\alias{pw_install}
\title{Install playwright in an already inited project}
\usage{
pw_install(where = golem::get_golem_wd())
}
\arguments{
\item{where}{The path to the package}
}
\description{
This function runs playwright init in a project
that has been previously initiated with \code{{pw}}
}
