% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_show_report.R
\name{pw_show_report}
\alias{pw_show_report}
\title{Show the Playwright report}
\usage{
pw_show_report(where = golem::get_golem_wd(), ...)
}
\arguments{
\item{where}{The path to the package}

\item{...}{Additional arguments to pass to \verb{npx playwright show-report}}
}
\description{
Show the Playwright report
}
