% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw-tests.R
\name{pw_test}
\alias{pw_test}
\title{Run Playwright tests}
\usage{
pw_test(where = golem::get_golem_wd(), ...)
}
\arguments{
\item{where}{The path to the package}

\item{...}{Additional arguments to pass to \verb{npx playwright test}}
}
\description{
Run Playwright tests
}
