% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw-use-github-action.R
\name{pw_use_github_action}
\alias{pw_use_github_action}
\title{Add the Gihub Action YAML for Playwright}
\usage{
pw_use_github_action()
}
\value{
Used for side effect
}
\description{
Add the Gihub Action YAML for Playwright
}
\examples{
if (FALSE){
  pw_use_github_action()
}
}
