% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw-use_test.R
\name{pw_use_test}
\alias{pw_use_test}
\title{Show the Playwright report}
\usage{
pw_use_test(name, where = golem::get_golem_wd())
}
\arguments{
\item{name}{The name of the test to add (without .test.ts extension)}

\item{where}{The path to the package}
}
\description{
Show the Playwright report
}
