% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_test_id.R
\name{with_test_id}
\alias{with_test_id}
\title{Add a data-testid attribute to a tag}
\usage{
with_test_id(tag, test_id)
}
\arguments{
\item{tag}{A shiny tag}

\item{test_id}{A string to use as the value of the data-testid attribute}
}
\value{
A shiny tag with the data-testid attribute added
}
\description{
This function adds a data-testid attribute to a shiny tag. This is useful for testing with \code{getByTestId} in Playwright.
}
