% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw-codegen.R
\name{pw_codegen}
\alias{pw_codegen}
\title{Run the code generator}
\usage{
pw_codegen(where = golem::get_golem_wd())
}
\arguments{
\item{where}{The root directory of your golem project}
}
\description{
This function runs the Playwright code generator.
}
