% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw-init.R
\name{pw_init}
\alias{pw_init}
\title{Initialize Playwright}
\usage{
pw_init(where = golem::get_golem_wd())
}
\arguments{
\item{where}{The path to the package}
}
\description{
This function initializes a Playwright project in the tests folder.
}
\examples{
\dontrun{
pw_init()
}
}
